Creation of AbstractModelFactory
