+++ org.adempiere.base/src/org/adempiere/base/DefaultModelFactory.java
@@ -17,9 +17,6 @@
  *****************************************************************************/
 package org.adempiere.base;
 
-import java.lang.reflect.Constructor;
-import java.sql.ResultSet;
-import java.util.Properties;
 import java.util.logging.Level;
 
 import org.compiere.model.MEntityType;
@@ -35,9 +32,9 @@ import org.compiere.util.Util;
  * @author Jorg Janke
  * @author hengsin
  */
-public class DefaultModelFactory implements IModelFactory {
+public class DefaultModelFactory extends AbstractModelFactory {
 
-	private static CCache<String,Class<?>> s_classCache = new CCache<String,Class<?>>(null, "PO_Class", 100, 120, false, 2000);
+	private CCache<String,Class<?>> s_classCache = new CCache<String,Class<?>>(null, "PO_Class", 100, 120, false, 2000);
 	private final static CLogger s_log = CLogger.getCLogger(DefaultModelFactory.class);
 
 	/**	Packages for Model Classes	*/
@@ -244,83 +241,5 @@ public class DefaultModelFactory implements IModelFactory {
 		if (s_log.isLoggable(Level.FINEST)) s_log.finest("Not found: " + className);
 		return null;
 	}	//	getPOclass
-
-	@Override
-	public PO getPO(String tableName, int Record_ID, String trxName) {
-		Class<?> clazz = getClass(tableName);
-		if (clazz == null)
-		{
-			return null;
-		}
-
-		boolean errorLogged = false;
-		try
-		{
-			Constructor<?> constructor = null;
-			try
-			{
-				constructor = clazz.getDeclaredConstructor(new Class[]{Properties.class, int.class, String.class});
-			}
-			catch (Exception e)
-			{
-				String msg = e.getMessage();
-				if (msg == null)
-					msg = e.toString();
-				s_log.warning("No transaction Constructor for " + clazz + " (" + msg + ")");
-			}
-
-			PO po = constructor!=null ? (PO)constructor.newInstance(new Object[] {Env.getCtx(), Integer.valueOf(Record_ID), trxName}) : null;
-			return po;
-		}
-		catch (Exception e)
-		{
-			if (e.getCause() != null)
-			{
-				Throwable t = e.getCause();
-				s_log.log(Level.SEVERE, "(id) - Table=" + tableName + ",Class=" + clazz, t);
-				errorLogged = true;
-				if (t instanceof Exception)
-					s_log.saveError("Error", (Exception)e.getCause());
-				else
-					s_log.saveError("Error", "Table=" + tableName + ",Class=" + clazz);
-			}
-			else
-			{
-				s_log.log(Level.SEVERE, "(id) - Table=" + tableName + ",Class=" + clazz, e);
-				errorLogged = true;
-				s_log.saveError("Error", "Table=" + tableName + ",Class=" + clazz);
-			}
-		}
-		if (!errorLogged)
-			s_log.log(Level.SEVERE, "(id) - Not found - Table=" + tableName
-				+ ", Record_ID=" + Record_ID);
-		return null;
-	}
-
-	@Override
-	public PO getPO(String tableName, ResultSet rs, String trxName) {
-		Class<?> clazz = getClass(tableName);
-		if (clazz == null)
-		{
-			return null;
-		}
-
-		boolean errorLogged = false;
-		try
-		{
-			Constructor<?> constructor = clazz.getDeclaredConstructor(new Class[]{Properties.class, ResultSet.class, String.class});
-			PO po = (PO)constructor.newInstance(new Object[] {Env.getCtx(), rs, trxName});
-			return po;
-		}
-		catch (Exception e)
-		{
-			s_log.log(Level.SEVERE, "(rs) - Table=" + tableName + ",Class=" + clazz, e);
-			errorLogged = true;
-			s_log.saveError("Error", "Table=" + tableName + ",Class=" + clazz);
-		}
-		if (!errorLogged)
-			s_log.log(Level.SEVERE, "(rs) - Not found - Table=" + tableName);
-		return null;
-	}
 
 }
